+++ lib/prelude.ml
@@ -16,7 +16,8 @@
 
 let env_and_file f =
   match Astring.String.cut ~sep:":" f with
-  | None -> (Block.Env.Default, f)
+  | None -> (Util.One_or_all.All, f)
   | Some (e, f) ->
-      if Astring.String.exists (( = ) ' ') e then (Block.Env.Default, f)
-      else (Block.Env.User_defined e, f)
+      if Astring.String.exists (( = ) ' ') e then
+        (Util.One_or_all.One Block.Env.Default, f)
+      else (Util.One_or_all.One (Block.Env.User_defined e), f)